requirements.txt
asgiref==3.11.0
Django==5.2.10
djangorestframework==3.16.1
et_xmlfile==2.0.0
numpy==2.4.1
openpyxl==3.1.5
pandas==3.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
xlrd==2.0.2
